features/usb: use CONFIG_NOP_USB_XCEIV as module
During our release tests we noticed that Apalis iMX8 was having some
boot freezes. After some investigations we discovered that it was
related to this 'CONFIG_NOP_USB_XCEIV' being enabled as builtin (while
on BSP, where the issue does not reproduce, it is a module).

So after changing this config to module on Torizon and trying to
retrigger this issue, we did not see any other boot freezes.

Related-to: TOR-3599

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>